+++ cpp/dolfinx/la/CMakeLists.txt
@@ -9,7 +9,6 @@ set(HEADERS_la
   ${CMAKE_CURRENT_SOURCE_DIR}/SparsityPattern.h
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Vector.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/VectorSpaceBasis.h
   PARENT_SCOPE)
 
 target_sources(dolfinx PRIVATE
